requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

docx2pdf==0.1.8
pandas==2.1.4
python-docx==1.1.0
